// payload-types.ts
export interface Task {
  id: number;
  title: string;
  description?: string | null;
  status?: ('todo' | 'in-progress' | 'done') | null;
  timeSpent?: number | null;
  updatedAt: string;
  createdAt: string;
}